requirements.txt
fastapi
uvicorn
python-dotenv
openai
langchain-openai
beautifulsoup4
requests
markdownify